MDL-75719 core_course: Fix completion criteria pills shown to user.
We change completion status to be COMPLETION_INCOMPLETE when user
get the grade, but not passing grade (assuming that completion criteria
is enabled and grade item is visible). But then it seems like user
has not started the activity. I.e. he will see 'Receive passing grade'
as todo, but really he should see it as failed. We still need to show
user that he failed that criteria if he has any grade.