MDL-75719 completion: Mark user completion as in progress.
Since we changed completion status to COMPLETION_INCOMPLETE
when user has not achieved passing grade (provided that completion
criteria is enabled and grade item is visible) we see course
progress as not started yet. So in this change we check if there is
any grade and if yes mark completion as in progress.

OT. mark_inprogress() method seems to be only used by mark_complete()
method before this change, which is a bit strange.